requirements.txt
windows-curses
python-dotenv
discord.py
Google-Images-Search